+++ base/accessory.js
@@ -42,7 +42,13 @@ class HomebridgeAccessory {
   }
 
   setDefaults () { 
-    this.config.allowResend = this.config.allowResend === undefined ? true : false; 
+    if(config.allowResend === undefined){
+      if(config.preventResendHex === undefined){
+        config.allowResend = true;
+      } else {
+        config.allowResend = !config.preventResendHex;
+      }
+    }
   }
 
   restoreStateOrder () { }
@@ -50,7 +56,7 @@ class HomebridgeAccessory {
   correctReloadedState () { }
 
   checkConfig (config) {
-    const { name, log } = this;
+    const { name, log, logLevel } = this;
     if (typeof config !== 'object') return;
 
     Object.keys(config).forEach((key) => {
@@ -149,7 +155,7 @@ class HomebridgeAccessory {
 
   async getCharacteristicValue (props, callback) {
     const { propertyName } = props;
-    const { log, name } = this;
+    const { log, name, logLevel } = this;
     let value;
 
     const capitalizedPropertyName = propertyName.charAt(0).toUpperCase() + propertyName.slice(1);
@@ -170,7 +176,7 @@ class HomebridgeAccessory {
   }
 
   loadState () {
-    const { config, log, name, serviceManager } = this;
+    const { config, log, logLevel, name, serviceManager } = this;
     let { host, resendDataAfterReload, resendDataAfterReloadDelay, persistState } = config;
 
     // Set defaults
@@ -253,7 +259,7 @@ class HomebridgeAccessory {
 
   // MQTT Support
   subscribeToMQTT () {
-    const { config, log, name } = this;
+    const { config, log, logLevel, name } = this;
     let { mqttTopic, mqttURL, mqttUsername, mqttPassword } = config;
 
     if (!mqttTopic || !mqttURL) return;
@@ -361,7 +367,7 @@ class HomebridgeAccessory {
   }
 
   mqttValueForIdentifier (identifier) {
-    const { log, name } = this;
+    const { log, logLevel, name } = this;
 
     let value = this.mqttValues[identifier];